FEATURE: Implement framework for user tutorial
This commit introduces a new framework for building user tutorials as
tips using the Tippy JS library. Currently, the new framework is used
to replace the old notification spotlight and tips and show a new one
related to the topic timeline.

Each new tip has a key that is used to store its state (active or
dismissed) in the database, in the user_options table. The existent
skip_new_user_tips key continues to exist as a key to store the global
state of all tips.

All tips follow the same structure and have a title, a description, an
optional image and two buttons for either dismissing just the current
tip or all of them at once. The first button simply updates the skip
key associated to the tip to "false" and the other button updates the
skip_new_user_tips user option to "true" and then all user tips are
marked as skipped.