Drop support for @action as string in DButton
Because glimmer components aren't provided a target context, they can't support `sendAction` like classic components. Therefore, DButton is no longer able to support an `@action` argument passed as a string.

On render, a deprecation warning is issued. If the action is triggered, via `click` for instance, then an error will be thrown.